request error handling and clean up
fixes #43
