+++ app/scripts/init.ts
@@ -1,3 +1,4 @@
 import './index'
 import './lightbox'
 import './logout'
+import './tabs'
